Move routes to new routers
