+++ package.json
@@ -40,7 +40,6 @@
     "github": "^13.1.0",
     "glob-to-regexp": "^0.3.0",
     "gravatar": "^1.8.0",
-    "highlight.js": "^9.14.2",
     "lodash": "^4.17.11",
     "markdown": "^0.5.0",
     "marked": "^0.7.0",
